Update signatures to support lazy Regulars
Functions that construct Regulars from other Regulars now also accept a
function that will later yield that a Regular on demand, and a function
`lazy` to turn such a function into a Regular is added. This makes it
possible to create circularly defined regulars, for rules like

X ::
    "x"
    "x" "," X

Note that they don’t actually work yet; some parsing algorithms have to
be adjusted in order to not immediately cause a stack overflow. This is
done in a separate commit.

Of course, this allows one to define non-regular languages, in contrast
to the name `ceylon.regular`. Users are advised to only use this to
define languages that are actually regular, such as the one above –

    "x" ("," "x")*

– and to use proper parsers, such as `ceylon.parse`, for irregular
languages.